refacoring where logVisibilityEvent code is and reverting keepalive setting for simplicity
