+++ packages/crashlytics/src/helpers.ts
@@ -15,11 +15,10 @@
  * limitations under the License.
  */
 
-import { SeverityNumber } from '@opentelemetry/api-logs';
+import { SeverityNumber, AnyValueMap } from '@opentelemetry/api-logs';
 import { CRASHLYTICS_TRACER_NAME } from './constants';
 import { Crashlytics } from './public-types';
 import { CrashlyticsInternal } from './types';
-import { logVisibilityEvent } from './api';
 import { type TimeInput } from '@opentelemetry/api';
 import { hrTimeToMilliseconds, timeInputToHrTime } from '@opentelemetry/core';
 
@@ -153,4 +152,33 @@ export function flush(crashlytics: Crashlytics): Promise<void> {
     .catch(err => {
       console.error('Error flushing logs from Firebase Crashlytics:', err);
     });
+}
+
+/**
+ * Logs a page visibility transition event (foreground or background).
+ *
+ * @public
+ *
+ * @param crashlytics - The {@link Crashlytics} instance.
+ * @param visibilityState - The current page visibility state ('visible' or 'hidden').
+ */
+export function logVisibilityEvent(
+  crashlytics: Crashlytics,
+  visibilityState: 'visible' | 'hidden'
+): void {
+  const { loggerProvider, attributesStore } =
+    crashlytics as CrashlyticsInternal;
+  const logger = loggerProvider.getLogger('visibility-logger');
+  const customAttributes: AnyValueMap = attributesStore.getLogAttributes();
+
+  const body =
+    visibilityState === 'hidden'
+      ? 'Background lifecycle event'
+      : 'Foreground lifecycle event';
+
+  logger.emit({
+    severityNumber: SeverityNumber.INFO,
+    body,
+    attributes: customAttributes
+  });
 }
